Text mode, still using DMA.
Here, instead of using a frame buffer, DMA alternates between
two line buffers. While one is being output, the other is being
prepared from a text mode screen and a text font.

An alternative would be to directly load SSI from interrupts
instead of using a buffer plus DMA. This was done because it was
a simpler change. Performance implications need to be
investigated.